refactor: moved scalar callbacks+event listeners to widgets.py
